add tensorflow and react-native-fs
however there is a weird bug with the metro config.

The tensorflow docs say that it's not required to change the metro
config however the app crashes with the following message:

"Error: Cannot evaluate flag 'DEBUG': no evaluation function found."

It looks like the app crashes because of the 'inlineRequires' flag
in the original metro config. More work needs to be done to understand
what witchcraft if happening over there